Добавлен контроллер резюме, изменена иерархия сервисов и небольшие изменения в аккаунте
