+++ src/main/java/com/example/hrplatform/repository/ResumeRepository.java
@@ -1,0 +1,9 @@
+package com.example.hrplatform.repository;
+
+import com.example.hrplatform.entity.Account;
+import com.example.hrplatform.entity.Resume;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ResumeRepository  extends JpaRepository<Resume, Long> {
+
+}
