+++ src/main/java/com/example/hrplatform/service/ResumeService.java
@@ -1,0 +1,7 @@
+package com.example.hrplatform.service;
+
+import com.example.hrplatform.entity.Resume;
+
+public interface ResumeService {
+    Resume addNewResume(Resume resume);
+}
